database: fix scan_dir
When probing for the type, I have made the classical mistake of using
as a parameter part of a structure that is moved into the capture. That
is what broke our tests.

But also, when stat'ing, de.name will give us only the component relative to
the current path. We need to add the directory so the stat will succeed.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>