storage_service: fix compilation on gcc 4.9
gcc 4.9 fails to compiles the following legal code (which compiles fine on
gcc 5.1):

	#include <unordered_set>
	std::unordered_set<int> hi() {
        	return {};
	}

Work around this problem to make Scylla compile on gcc 4.9 again.

Note that this bug is specific to std::unordered_set - we have other places
in the code code which "return {}" for std::set, and those work fine.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>