version: change all fields to uint16_t
Ok, shame on me: the version string was so obviously correct that I only
verified that the comparisons were working as expected.

Turns out it isn't: http://lists.boost.org/boost-users/2006/12/24194.php

boost::format will treat uint8_t arguments as char, and therefore we will end
up with the version string misprinted.

We can just cast it to uint16_t before we print, but since this is not exactly
a struct that we will be using all the time, let's favor readability over
saving a few bytes, and change all fields to uint16_t.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>