row consumer: extend use of read for multi-value fields
In an attempt to gain some cycles, we are testing whether we can read many
values at once, and if so, using consume_be directly for those.

What we can do in this situation, is read the first value, and let the read
fall through the next case if the read succeeds.

The code actually looks a lot more elegant this way.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>