transport: protect against race in using cql connection gate
While we have a with_gate() call when using the connection, it is
somewhere deep in the bowels of process_request_one().  If that
continuation is deferred, the gate can be closed without anyone
noticing that a request is still working on it; when we do hit the
with_gate, we'll see a use-after-free.