sstable: move buffer size inside the sstable object.
We'll pay the price of having this now as a variable instead of a constexpr,
but this dims in comparison with the rest of the operation.

By paying this cost, we gain the ability of actually specifying it during test
runs, making it easy to automate scripts that will measure the performance over
various buffer sizes.

I am also providing a new constructor that allows for the setting of the buffer
size.  The said constructor will be private, meaning that only the test class
will be able to use it.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>