sstables: don't return after processing collections
The code as is is blatantly wrong, and is an artifact of the seastar-thread
conversion.

This happened because the way we move to the next element in a do_for_each
future loop, is by returning the current lambda, and so it was converted this
way. Since we are now using a for loop, we should not return: we should continue.

I found this while searching for a bug, which is unfortunately not fixed by this.
But this is totally wrong, and has to be fixed.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>