column helper: loop once only while gathering statistics.
the code to gather statistics about the column_name is showing in the
benchmark.

If we really want to collect those statistics, I guess they will never be free
because they involve a byte copy which implies an allocation.

But one easy thing we can do to make it better, is collect both min and max
statistics in the same loop. There is also no need to special case the case of
an empty vector, since may_grow will already take care of that.

That yields us a ~ 0.77 % boost, which although not earth shattering, is easy
enough for us not to reap.

Before:
200582.94 +- 293.91 partitions / sec (30 runs, 1 concurrent ops)

After:
202120.06 +- 341.95 partitions / sec (30 runs, 1 concurrent ops)

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>