sstables::mutation_reader: do not spill previous collection
The sstable mutation_reader knows how to handle the case of multiple
mutations: it wraps the mutation into an optional element, and when a
new mutation comes along, it just moves that to the caller.

However, it fails to do the same for collections. We have the same collection
tracking structure, and as long as we are trying to read a single row, all
works fine.

When we are reading more than a row, though, things change: because we have
stale data for the previous collection, the new one will end up including this.

In the case in which we tested, we had an sstable that contained mutations for
all columns in a row, and another row containing all columns but one, and that
one was precisely a collection. What we ended up seeing, is that the new row
would have that collection nevertheless, with the former row's collection
value.

This patch applies the same method as we use for tracking mutations, to tracking
collections. And with this, it all works.

Fixes #188

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>