storage_proxy: log failed attempts at communicating
If an exception happens in the query path, we'll never know about it. They are
currently being ignored.

Investigating this, I found out that this is because the readers in
storage_proxy.cc handles them - but don't log they anywhere.

This patch introduces such logging. the error() function takes an sstring not
an exception_ptr: this is so we can reuse it in the future to also log problems
from other hosts (currently not done).

We have a separate helper to extract the message from the current exception
before we pass it to error()

Fixes #110

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>