sstables: fix shared sstable deletion
If an sstable is shared, we can't just go ahead and remove it, because
other shards may still be using it; we need their agreement.

Since each shard will have its own sstable object for the same sstable,
we can't use the sstable class as a synchronization point.  Instead,
use a static unordered map indexed by the TOC file name.