storage_service: return values for exception, native_transport and rpc_server
This patch changes the behaviour of is_native_transport_running and
is_rpc_server_running to return true and not to fail, we assume that
they are running. It should be changed when an API to start and stop
them will be added.

The get_exception_count will return 0, the definition for it in origin
is exception that were not cought in a thread.

We should re-think about what it means in our implementation, meanwhile
return 0, for no exception, is a reasonable approach.

Signed-off-by: Amnon Heiman <amnon@cloudius-systems.com>