build: configure seastar compiler with right -march flag
We set -march=nehalem to get a cross-microarch binary that supports the
crc32 instruction, but we don't configure seastar with that flag; so seastar
instead inherits dpdk's default, which is -march=native.  The resulting
binary may not run on hosts other than the one it was compiled on.

Fix by passing the flag to seastar and inheriting it from there.

Fixes #338.